requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
